Source: autoclone repo if do not exist

+++ c_source-sync.sh
@@ -7,12 +7,21 @@ source $MYFOLDER/colors.sh
 if [ ! -d "$HOME/apps" ]; then
   mkdir ~/apps
 fi
-cd ~/apps/android-vts/
-echo -e ${GREEN}".o0o. .o0o..o0o..o0o. Patching $PWD .o0o. .o0o..o0o..o0o."${NC}
-git reset --hard
-git clean -fxd :/
-git fetch origin master
-git checkout origin/master
+
+if [ ! -d "$HOME/apps/android-vts/.git" ]; then
+  echo -e ${RED}".o0o. .o0o..o0o..o0o. $HOME/apps/android-vts/.git do not exist! .o0o. .o0o..o0o..o0o. "${NC}
+  echo -e ${RED}".o0o. .o0o..o0o..o0o. Cloning! .o0o. .o0o..o0o..o0o. "${NC}
+  cd ~/apps
+  git clone https://github.com/AndroidVTS/android-vts -b master android-vts
+  cd ~/apps/android-vts/
+else 
+  cd ~/apps/android-vts/
+  echo -e ${GREEN}".o0o. .o0o..o0o..o0o. Patching $PWD .o0o. .o0o..o0o..o0o."${NC}
+  git reset --hard
+  git clean -fxd :/
+  git fetch origin master
+  git checkout origin/master
+fi
 
 wget "https://raw.githubusercontent.com/nailyk-fr/jenkins_scripts/nougat-mr1.1-release/patches/AndroidVTS.properties" -O local.properties
 cat "/var/jenkins_home/workspace/F-Droid/.signing" >> app/gradle.properties
@@ -34,13 +43,20 @@ android {
     }
 }" >> app/build.gradle
 
-
-cd ~/apps/mGerrit/
-echo -e ${GREEN}".o0o. .o0o..o0o..o0o. Patching $PWD .o0o. .o0o..o0o..o0o."${NC}
-git reset --hard
-git clean -fxd :/
-git fetch origin master
-git checkout origin/master
+if [ ! -d "$HOME/apps/mGerrit/.git" ]; then
+  echo -e ${RED}".o0o. .o0o..o0o..o0o. $HOME/apps/mGerrit/.git do not exist! .o0o. .o0o..o0o..o0o. "${NC}
+  echo -e ${RED}".o0o. .o0o..o0o..o0o. Cloning! .o0o. .o0o..o0o..o0o. "${NC}
+  cd ~/apps
+  git clone https://github.com/JBirdVegas/external_jbirdvegas_mGerrit -b master mGerrit
+  cd ~/apps/mGerrit/
+else 
+  cd ~/apps/mGerrit/
+  echo -e ${GREEN}".o0o. .o0o..o0o..o0o. Patching $PWD .o0o. .o0o..o0o..o0o."${NC}
+  git reset --hard
+  git clean -fxd :/
+  git fetch origin master
+  git checkout origin/master
+fi
 
 wget "https://raw.githubusercontent.com/nailyk-fr/jenkins_scripts/nougat-mr1.1-release/patches/AndroidVTS.properties" -O local.properties
 cat "/var/jenkins_home/workspace/F-Droid/.signing" >> app/gradle.properties
@@ -114,13 +130,55 @@ index 7de304f..adcd575 100644
 " | patch -p1
 rm app/src/main/res/values/arrays.xml.orig || true
 
+if [ ! -d "$HOME/apps/Tusky/.git" ]; then
+  echo -e ${RED}".o0o. .o0o..o0o..o0o. $HOME/apps/Tusky/.git do not exist! .o0o. .o0o..o0o..o0o. "${NC}
+  echo -e ${RED}".o0o. .o0o..o0o..o0o. Cloning! .o0o. .o0o..o0o..o0o. "${NC}
+  cd ~/apps
+  git clone https://github.com/tuskyapp/Tusky -b master Tusky
+  cd ~/apps/Tusky/
+else 
+  cd ~/apps/Tusky/
+  echo -e ${GREEN}".o0o. .o0o..o0o..o0o. Patching $PWD .o0o. .o0o..o0o..o0o."${NC}
+  git reset --hard
+  git clean -fxd :/
+  git fetch origin master
+  git checkout origin/master
+fi
 
-cd ~/apps/Tusky/
-echo -e ${GREEN}".o0o. .o0o..o0o..o0o. Patching $PWD .o0o. .o0o..o0o..o0o."${NC}
-git reset --hard
-git clean -fxd :/
-git fetch origin master
-git checkout origin/master
+wget "https://raw.githubusercontent.com/nailyk-fr/jenkins_scripts/nougat-mr1.1-release/patches/AndroidVTS.properties" -O local.properties
+cat "/var/jenkins_home/workspace/F-Droid/.signing" >> app/gradle.properties
+echo "
+android {
+    signingConfigs {
+       release {
+           storeFile file(RELEASE_STORE_FILE)
+           storePassword RELEASE_STORE_PASSWORD
+           keyAlias RELEASE_KEY_ALIAS
+           keyPassword RELEASE_KEY_PASSWORD
+       }
+    }
+
+    buildTypes {
+            release {
+                signingConfig signingConfigs.release
+            }
+    }
+}" >> app/build.gradle
+
+if [ ! -d "$HOME/apps/OpenLauncher/.git" ]; then
+  echo -e ${RED}".o0o. .o0o..o0o..o0o. $HOME/apps/OpenLauncher/.git do not exist! .o0o. .o0o..o0o..o0o. "${NC}
+  echo -e ${RED}".o0o. .o0o..o0o..o0o. Cloning! .o0o. .o0o..o0o..o0o. "${NC}
+  cd ~/apps
+  git clone https://github.com/OpenLauncherTeam/openlauncher -b master OpenLauncher
+  cd ~/apps/OpenLauncher/
+else 
+  cd ~/apps/OpenLauncher/
+  echo -e ${GREEN}".o0o. .o0o..o0o..o0o. Patching $PWD .o0o. .o0o..o0o..o0o."${NC}
+  git reset --hard
+  git clean -fxd :/
+  git fetch origin master
+  git checkout origin/master
+fi
 
 wget "https://raw.githubusercontent.com/nailyk-fr/jenkins_scripts/nougat-mr1.1-release/patches/AndroidVTS.properties" -O local.properties
 cat "/var/jenkins_home/workspace/F-Droid/.signing" >> app/gradle.properties
@@ -143,12 +201,20 @@ android {
 }" >> app/build.gradle
 
 
-cd ~/apps/OpenLauncher/
-echo -e ${GREEN}".o0o. .o0o..o0o..o0o. Patching $PWD .o0o. .o0o..o0o..o0o."${NC}
-git reset --hard
-git clean -fxd :/
-git fetch origin master
-git checkout origin/master
+if [ ! -d "$HOME/apps/Osmand/.git" ]; then
+  echo -e ${RED}".o0o. .o0o..o0o..o0o. $HOME/apps/Osmand/.git do not exist! .o0o. .o0o..o0o..o0o. "${NC}
+  echo -e ${RED}".o0o. .o0o..o0o..o0o. Cloning! .o0o. .o0o..o0o..o0o. "${NC}
+  cd ~/apps
+  git clone https://github.com/osmandapp/Osmand -b master Osmand
+  cd ~/apps/Osmand/
+else 
+  cd ~/apps/Osmand
+  echo -e ${GREEN}".o0o. .o0o..o0o..o0o. Patching $PWD .o0o. .o0o..o0o..o0o."${NC}
+  git reset --hard
+  git clean -fxd :/
+  git fetch origin master
+  git checkout origin/master
+fi
 
 wget "https://raw.githubusercontent.com/nailyk-fr/jenkins_scripts/nougat-mr1.1-release/patches/AndroidVTS.properties" -O local.properties
 cat "/var/jenkins_home/workspace/F-Droid/.signing" >> app/gradle.properties
@@ -170,43 +236,23 @@ android {
     }
 }" >> app/build.gradle
 
+if [ ! -d "$HOME/apps/vlc-android/.git" ]; then
+  echo -e ${RED}".o0o. .o0o..o0o..o0o. $HOME/apps/vlc-android/.git do not exist! .o0o. .o0o..o0o..o0o. "${NC}
+  echo -e ${RED}".o0o. .o0o..o0o..o0o. Cloning! .o0o. .o0o..o0o..o0o. "${NC}
+  cd ~/apps
+  git clone https://code.videolan.org/videolan/vlc-android -b master vlc-android
+  cd ~/apps/vlc-android/
+else 
+  cd ~/apps/vlc-android/
+  echo -e ${GREEN}".o0o. .o0o..o0o..o0o. Patching $PWD .o0o. .o0o..o0o..o0o."${NC}
+  git reset --hard
+  git clean -fxd :/
+  git fetch origin master
+  git checkout origin/master
+fi
 
-cd ~/apps/Osmand
-echo -e ${GREEN}".o0o. .o0o..o0o..o0o. Patching $PWD .o0o. .o0o..o0o..o0o."${NC}
-git reset --hard
-git clean -fxd :/
-git fetch origin master
-git checkout origin/master
-wget "https://raw.githubusercontent.com/nailyk-fr/jenkins_scripts/nougat-mr1.1-release/patches/AndroidVTS.properties" -O local.properties
-cat "/var/jenkins_home/workspace/F-Droid/.signing" >> app/gradle.properties
-echo "
-android {
-    signingConfigs {
-       release {
-           storeFile file(RELEASE_STORE_FILE)
-           storePassword RELEASE_STORE_PASSWORD
-           keyAlias RELEASE_KEY_ALIAS
-           keyPassword RELEASE_KEY_PASSWORD
-       }
-    }
-
-    buildTypes {
-            release {
-                signingConfig signingConfigs.release
-            }
-    }
-}" >> app/build.gradle
-
-
-cd ~/apps/vlc-android/
-echo -e ${GREEN}".o0o. .o0o..o0o..o0o. Patching $PWD .o0o. .o0o..o0o..o0o."${NC}
-git reset --hard
-git clean -fxd :/
-git fetch origin master
-git checkout origin/master
 cp "/var/jenkins_home/workspace/F-Droid/.signing_vlc" gradle.properties
 wget "https://raw.githubusercontent.com/nailyk-fr/jenkins_scripts/nougat-mr1.1-release/patches/AndroidVTS.properties" -O local.properties
-
 
 cd ~/apps/vlc-android/vlc
 git reset --hard
